Updated Secuity across shelf functions

+++ AutoBuildApp/AutoBuildApp_v3/Backend/AutoBuildApp.Services/FeatureServices/ShelfWithResponseServices.cs
@@ -5,6 +5,8 @@ using AutoBuildApp.Models.DataTransferObjects;
 using AutoBuildApp.DataAccess;
 using AutoBuildApp.Models.Products;
 using AutoBuildApp.Models.Enumerations;
+using AutoBuildApp.Security.Enumerations;
+using AutoBuildApp.Security;
 
 /**
 * The ShelfWithResponseService will provide the shelf specific operatons
@@ -15,10 +17,19 @@ namespace AutoBuildApp.Services
 {
     public class ShelfWithResponseService
     {
-        private ShelfDAO _dao;
+        private readonly ShelfDAO _dao;
+        private readonly List<string> _approvedRoles;
 
         public ShelfWithResponseService(ShelfDAO shelfDAO)
         {
+            _approvedRoles = new List<string>()
+            {
+                RoleEnumType.BasicRole,
+                RoleEnumType.DelegateAdmin,
+                RoleEnumType.VendorRole,
+                RoleEnumType.SystemAdmin
+            };
+
             _dao = shelfDAO;
         }
 
@@ -31,6 +42,14 @@ namespace AutoBuildApp.Services
         public CommonResponse CreateShelf(string shelfName, string username)
         {
             CommonResponse output = new CommonResponse();
+
+            if (!AuthorizationCheck.IsAuthorized(_approvedRoles))
+            {
+                output.ResponseString = ResponseStringGlobals.UNAUTHORIZED_ACCESS;
+                output.ResponseBool = false;
+                return output;
+            }
+
             var temp = _dao.InsertShelf(shelfName, username);
 
             output.ResponseBool = temp.GenericObject;
@@ -57,6 +76,14 @@ namespace AutoBuildApp.Services
         public CommonResponse DeleteShelf(string shelfName, string username)
         {
             CommonResponse output = new CommonResponse();
+
+            if (!AuthorizationCheck.IsAuthorized(_approvedRoles))
+            {
+                output.ResponseString = ResponseStringGlobals.UNAUTHORIZED_ACCESS;
+                output.ResponseBool = false;
+                return output;
+            }
+
             var temp = _dao.DeleteShelf(shelfName, username);
 
             output.ResponseBool = temp.GenericObject;
@@ -84,6 +111,14 @@ namespace AutoBuildApp.Services
         public CommonResponse ChangeShelfName(string oldName, string newName, string username)
         {
             CommonResponse output = new CommonResponse();
+
+            if (!AuthorizationCheck.IsAuthorized(_approvedRoles))
+            {
+                output.ResponseString = ResponseStringGlobals.UNAUTHORIZED_ACCESS;
+                output.ResponseBool = false;
+                return output;
+            }
+
             var temp = _dao.UpdateShelfName(oldName, newName, username);
 
             output.ResponseBool = temp.GenericObject;
@@ -112,6 +147,14 @@ namespace AutoBuildApp.Services
         public CommonResponse AddToShelf(string modelNumber, int quantity, string shelfName, string username)
         {
             CommonResponse output = new CommonResponse();
+
+            if (!AuthorizationCheck.IsAuthorized(_approvedRoles))
+            {
+                output.ResponseString = ResponseStringGlobals.UNAUTHORIZED_ACCESS;
+                output.ResponseBool = false;
+                return output;
+            }
+
             var temp = _dao.AddComponent(modelNumber, quantity, shelfName, username);
 
             output.ResponseBool = temp.GenericObject;
@@ -139,6 +182,14 @@ namespace AutoBuildApp.Services
         public CommonResponse RemoveFromShelf(int itemIndex, string shelfName, string username)
         {
             CommonResponse output = new CommonResponse();
+
+            if (!AuthorizationCheck.IsAuthorized(_approvedRoles))
+            {
+                output.ResponseString = ResponseStringGlobals.UNAUTHORIZED_ACCESS;
+                output.ResponseBool = false;
+                return output;
+            }
+
             var temp = _dao.RemoveComponent(itemIndex, shelfName, username);
 
             output.ResponseBool = temp.GenericObject;
@@ -166,6 +217,14 @@ namespace AutoBuildApp.Services
         public CommonResponse ReorderShelf(List<int> indicies, string shelfName, string username)
         {
             CommonResponse output = new CommonResponse();
+
+            if (!AuthorizationCheck.IsAuthorized(_approvedRoles))
+            {
+                output.ResponseString = ResponseStringGlobals.UNAUTHORIZED_ACCESS;
+                output.ResponseBool = false;
+                return output;
+            }
+
             var temp = _dao.UpdateShelfOrder(indicies, shelfName, username);
 
             output.ResponseBool = temp.GenericObject;
@@ -191,9 +250,17 @@ namespace AutoBuildApp.Services
         /// <param name="shelfName">The name of the shelf the item resides on.</param>
         /// <param name="username">The name of the user the shelf is associated with.</param>
         /// <returns>True or false with a response string.</returns>
-        public CommonResponse ChangeQuantity(int itemIndex, int quantity, string shelfName, string username)
+        public CommonResponse UpdateQuantity(int itemIndex, int quantity, string shelfName, string username)
         {
             CommonResponse output = new CommonResponse();
+
+            if (!AuthorizationCheck.IsAuthorized(_approvedRoles))
+            {
+                output.ResponseString = ResponseStringGlobals.UNAUTHORIZED_ACCESS;
+                output.ResponseBool = false;
+                return output;
+            }
+
             var temp = _dao.UpdateQuantity(itemIndex, quantity, shelfName, username);
 
             output.ResponseBool = temp.GenericObject;
@@ -223,6 +290,14 @@ namespace AutoBuildApp.Services
             {
                 ResponseBool = false
             };
+
+            if (!AuthorizationCheck.IsAuthorized(_approvedRoles))
+            {
+                output.GenericObject = new Shelf();
+                output.ResponseString = ResponseStringGlobals.UNAUTHORIZED_ACCESS;
+                return output;
+            }
+
             var temp = _dao.GetShelfByName(shelfName, username);
             output.GenericObject = temp.GenericObject;
 
@@ -241,13 +316,21 @@ namespace AutoBuildApp.Services
         /// </summary>
         /// <param name="username">Name associated with the account to retrieve from.</param>
         /// <returns></returns>
-        public CommonResponseWithObject<List<Shelf>> GetAllusernameShelves(string username)
+        public CommonResponseWithObject<List<Shelf>> GetShelvesByUser(string username)
         {
             CommonResponseWithObject<List<Shelf>> output = new CommonResponseWithObject<List<Shelf>>()
             {
                 ResponseBool = false
             };
-            var temp = _dao.GetAllShelvesByUser(username);
+
+            if (!AuthorizationCheck.IsAuthorized(_approvedRoles))
+            {
+                output.GenericObject = new List<Shelf>();
+                output.ResponseString = ResponseStringGlobals.UNAUTHORIZED_ACCESS;
+                return output;
+            }
+
+            var temp = _dao.GetShelvesByUser(username);
             output.GenericObject = temp.GenericObject;
 
             CodeToStringHandler(output, temp.Code);
@@ -268,9 +351,9 @@ namespace AutoBuildApp.Services
         //    return output;
         //}
 
+        #region Private Helper Methods
         /// <summary>
         /// Helper function to insert the strings from each expected error code.
-        #region Private Helper Method
         /// </summary>
         private void CodeToStringHandler(CommonResponse commonResponse, AutoBuildSystemCodes code)
         {
